Dag Works!
Workload shaping "works" but feels pretty hard coded right now.
It's possible to hit the targeted bug yet again but the dag makes it a
lot easier to adjust workloads and add new elements.
Still feels like we'd want to codegen a lot of the stuff here.

Need to work on serializing transcripts.
Having a good DAG diff is going to be pretty important. Could note state
and then the expected diff to save on space.
Given the nature of this DAG, IDs could likely be built as a hash of the
content + a hash of the edges? Or alternatively, since everything is
namespaced, performing a traversal back to the nearest database would
generate a fairly stable "ID".... I like that.